README: Change robot initial pose to multiples of pi/2
This is important so that the move_base costmap is aligned with the
global costmap. This is always the case on the real robot independent of
the startup position, but on Gazebo you should take care to always use
multiples of pi/2. Otherwise, there will be discretization artifacts due
to the diagonal lines in the costmap.