Add klt parameter to store_item
Note: we still need to add a new in_box fluent.
Until that is done it will create an insufficient plan
that adds the multimeter to klt_2 once it
sees it, then leaves it there and only searches for klt_1.
This is because of the currently used in_box fluent.
Therefore, also the goal needs to be adapted.