Delete solve() method of TablesDemoDomain
Instead of that it uses the solve-method in the ESB's Domain
which returns a plan instead of a list of actions.
This makes it more re-usable for applications where we need
the full plan instead of the simple list of actions.
This is for example the case when we use the ESB's dispatcher.